case file link modal moved to case files folder to reuse
